feat: halloween functioning lights

+++ main/include/led.h
@@ -26,7 +26,7 @@ typedef struct {
     led_t *leds;
     uint16_t led_count;
     uint8_t start_universe;
-    uint8_t start_address;
+    uint16_t start_address;
 } led_strip_t;
 
 void initialize_led(led_t *led, uint16_t id, uint8_t universe, uint16_t dmx_address);
